Add Swets
Added Swets.
Right now they use the slime models and textures as placeholders until custom models and textures are created.
They do everything that they did in the original aether legacy.
This implementation allows for more swet variants if that should be wanted in the future.
Things that need to be changed: swet loot tables and spawning.