Change sigmoid contrast curve
The original curve was very expensive, with 3 calls to Math.exp()
The new method is much less expensive, with just one call to Math.sqrt.
The shape of the sigmoid also better lends to less unnatural leaf colorings, or so I believe.
The contrast values have been adjusted to match as closely as possible to the original curve.