Fluid rendering changes
Still water will now have the same fading effects as flowing water in the Aether.
When swimming in dense aercloud, there's now an fitting overlay such that it doesn't look so odd. At the moment, the overlay is the dense aercloud texture, but probably should be changed to something custom.
I couldn't figure out how to remove the effect that makes the screen blue, so that remains.