Various FloatingBlockEntity changes
Removes an obsolete and unused method in FloatingBlock.
onLand has been renamed to onEndFloating, and now executes for blocks that don't land. (AKA TNT will explode whether or not it breaks on a slab.) Because of changes to the Biconsumer, landing effects can differ based on whether the block properly lands. For example, lightning rods will not summon lightning if they break on a slab.